+++ src/dc/SeedRound.cpp
@@ -78,7 +78,7 @@ std::unique_ptr<DCState> SeedRound::executeTask() {
 
         // initialize the slices of the k-th share with zeroes
         // except the slices of the own message slot
-        if (slotIndex_ == slot) {
+        if (static_cast<uint32_t>(slotIndex_) == slot) {
             for (uint32_t slice = 0; slice < numSlices; slice++)
                 shares[slot][k_ - 1].push_back(messageSlices[slice]);
         } else {
